+++ Public/Files/Add-OpenAIFile.ps1
@@ -16,7 +16,7 @@ function Add-OpenAIFile {
         [string]$Name,
 
         [Parameter(Mandatory)]
-        [Completions('assistants', 'fine-tune', 'batch')]
+        [Completions('assistants', 'vision', 'fine-tune', 'batch')]
         [ValidateNotNullOrEmpty()]
         [string][LowerCaseTransformation()]$Purpose,
 
@@ -75,12 +75,13 @@ function Add-OpenAIFile {
         #region Construct parameters for API request
         $PostBody = [System.Collections.Specialized.OrderedDictionary]::new()
         if ($PSCmdlet.ParameterSetName -eq 'File') {
-            if (-not $File.Exists) {
+            [string]$refT = ''
+            if (-not $File.Exists -and -not $PSCmdlet.SessionState.Path.IsPSAbsolute($File.FullName, [ref]$refT)) {
                 # Try to resolve relative path
                 $File = [System.IO.FileInfo]$PSCmdlet.SessionState.Path.GetUnresolvedProviderPathFromPSPath($File.Name)
             }
             if (-not $File.Exists) {
-                Write-Error -Exception ([System.Management.Automation.ItemNotFoundException]::new(("Cannot find path '{0}' because it does not exist." -f $File)))
+                Write-Error -Exception ([System.Management.Automation.ItemNotFoundException]::new(("Cannot find path '{0}' because it does not exist." -f $File.FullName)))
                 return
             }
             $PostBody.file = $File