+++ Tests/Threads.E2E.tests.ps1
@@ -176,5 +176,85 @@ Password should be between 8 and 12 random alphanumeric characters.
                 { $script:UploadItems | Remove-OpenAIFile -ea Stop } | Should -Not -Throw
             }
         }
+
+        Context 'Vision' {
+            BeforeAll {
+                $script:Assistant = $null
+                $script:Thread = $null
+
+                # sample image url
+                $script:WikiDonutsImageUrl = 'https://upload.wikimedia.org/wikipedia/commons/thumb/e/e4/4_donuts.jpg/640px-4_donuts.jpg'
+
+                $script:Instructions = 'You are an expert in pastries'
+                $script:PromptMessage = 'Please explain the relationship between the two images.'
+            }
+
+            It 'STEP1: Uploads test image' {
+                { $script:UploadItems = Add-OpenAIFile -File ($script:TestData + '/sweets_donut.png') -Purpose vision -ea Stop } | Should -Not -Throw
+                $UploadItems.id | Should -BeLike 'file*'
+                $UploadItems.purpose | Should -Be 'vision'
+            }
+
+            It 'STEP2: Create an Assistant' {
+                $RandomName = ('TEST' + (Get-Random -Maximum 1000))
+                { $splat = @{
+                        Name          = $RandomName
+                        Model         = 'gpt-4-turbo'
+                        Description   = 'Test assistant'
+                        Instructions  = $script:Instructions
+                        Temperature   = 0
+                        TimeoutSec    = 30
+                        MaxRetryCount = 5
+                        ErrorAction   = 'Stop'
+                    }
+                    $script:Assistant = New-Assistant @splat
+                } | Should -Not -Throw
+                $Assistant.id | Should -BeLike 'asst_*'
+                $Assistant.object | Should -BeExactly 'assistant'
+                $Assistant.name | Should -Be $RandomName
+            }
+
+            It 'STEP3: Create a Thread and Add a message' {
+                { $script:Thread = New-Thread -TimeoutSec 30 -MaxRetryCount 5 -ea Stop | `
+                            Add-ThreadMessage `
+                            -Message $script:PromptMessage `
+                            -Images @($script:UploadItems, $script:WikiDonutsImageUrl) `
+                            -PassThru -TimeoutSec 30 -MaxRetryCount 5 -ea Stop } | Should -Not -Throw
+                $Thread.id | Should -BeLike 'thread_*'
+                $Thread.Messages | Should -HaveCount 1
+            }
+
+            It 'STEP4: Start Thread Run, then wait for run completion and get result.' {
+                { $script:Thread = $script:Thread | Start-ThreadRun -Assistant $script:Assistant -ea Stop |
+                        Receive-ThreadRun -Wait -TimeoutSec 100 -ea Stop } | Should -Not -Throw
+                $Thread.id | Should -BeLike 'thread_*'
+                $Thread.Messages.Count | Should -BeGreaterOrEqual 2
+            }
+
+            It 'STEP5: Get result messages as simple and read suitable format.' {
+                $Thread.Messages.SimpleContent[0] | Should -BeOfType [pscustomobject]
+                $Thread.Messages.SimpleContent[0].Role | Should -Be 'user'
+                $Thread.Messages.SimpleContent[0].Type | Should -Be 'text'
+                $Thread.Messages.SimpleContent[0].Content | Should -BeExactly $script:PromptMessage
+                $Thread.Messages.SimpleContent[-1] | Should -BeOfType [pscustomobject]
+                $Thread.Messages.SimpleContent[-1].Role | Should -Be 'assistant'
+                $Thread.Messages.SimpleContent[-1].Type | Should -Be 'text'
+            }
+
+            It 'STEP6: Get run step details' {
+                { $script:Steps = $script:Thread | Get-ThreadRun -All -ea Stop | Get-ThreadRunStep -All -ea Stop } | Should -Not -Throw
+                $Steps.Count | Should -BeGreaterOrEqual 1
+                $Steps.SimpleContent[-1] | Should -BeOfType [pscustomobject]
+                $Steps.SimpleContent.Role | Should -Contain 'assistant'
+                $Steps.SimpleContent.Type | Should -Contain 'text'
+                $Steps.SimpleContent[-1].Content | Should -BeOfType [string]
+            }
+
+            It 'STEP7: Cleanup' {
+                { $script:Thread | Remove-Thread -ea Stop } | Should -Not -Throw
+                { $script:Assistant | Remove-Assistant -ea Stop } | Should -Not -Throw
+                { $script:UploadItems | Remove-OpenAIFile -ea Stop } | Should -Not -Throw
+            }
+        }
     }
 }