+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ### Unreleased
+- Add `-Images` parameter to `Add-ThreadMessage`.  
+  This makes that the image input can be used in Assistants.
 - Rename `Register-OpenAIFile` to `Add-OpenAIFile`.  
   (Note: You can still use `Register-OpenAIFile` as an alias. However, in the future release, the alias will be removed.)